[FIX] 패키지 이름 수정 및 로그인 로직 조정
