+++ src/main/java/com/rhkr8521/spaceon/api/member/dto/KakaoLoginRequestDTO.java
@@ -1,0 +1,8 @@
+package com.rhkr8521.spaceon.api.member.dto;
+
+import lombok.Data;
+
+@Data
+public class KakaoLoginRequestDTO {
+    private String accessToken;
+}
