+++ src/main/java/com/rhkr8521/spaceon/api/member/dto/KakaoUserInfoDTO.java
@@ -1,0 +1,10 @@
+package com.rhkr8521.spaceon.api.member.dto;
+
+import lombok.Data;
+
+@Data
+public class KakaoUserInfoDTO {
+    private String id;            // 사용자 고유 ID
+    private String email;         // 이메일
+    private String nickname;      // 닉네임
+}
