+++ src/main/java/com/rhkr8521/spaceon/api/member/repository/MemberRepository.java
@@ -1,0 +1,16 @@
+package com.rhkr8521.spaceon.api.member.repository;
+
+import com.rhkr8521.spaceon.api.member.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+    Optional<Member> findByEmail(String email);
+
+    Optional<Member> findByRefreshToken(String refreshToken);
+
+    Optional<Member> findBySocialId(String socialId);
+
+}
