+++ src/main/java/com/rhkr8521/spaceon/api/member/service/MemberService.java
@@ -1,0 +1,62 @@
+package com.rhkr8521.spaceon.api.member.service;
+
+import com.rhkr8521.spaceon.api.member.dto.*;
+import com.rhkr8521.spaceon.api.member.entity.*;
+import com.rhkr8521.spaceon.api.member.jwt.service.JwtService;
+import com.rhkr8521.spaceon.api.member.repository.MemberRepository;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.*;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class MemberService {
+
+    private final MemberRepository memberRepository;
+    private final JwtService jwtService;
+    private final OAuthService oAuthService;
+
+    @Transactional
+    public Map<String, Object> loginWithKakao(String kakaoAccessToken) {
+        // 카카오 Access Token을 이용해 사용자 정보 가져오기
+        KakaoUserInfoDTO kakaoUserInfo = oAuthService.getKakaoUserInfo(kakaoAccessToken);
+
+        // 사용자 정보를 저장
+        Member member = registerOrLoginKakaoUser(kakaoUserInfo);
+
+        // 액세스, 리프레시 토큰 생성
+        Map<String, String> tokens = jwtService.createAccessAndRefreshToken(member.getEmail());
+
+        Map<String, Object> response = new HashMap<>();
+        response.put("tokens", tokens);
+        response.put("profileImage", member.getProfileImage());
+
+        return response;
+    }
+
+    // 카카오 사용자 정보를 사용해 회원가입 또는 로그인 처리
+    public Member registerOrLoginKakaoUser(KakaoUserInfoDTO kakaoUserInfo) {
+        // 카카오 사용자 ID로 사용자 조회
+        return memberRepository.findBySocialId(kakaoUserInfo.getId())
+                .orElseGet(() -> registerNewKakaoUser(kakaoUserInfo));  // 없으면 새 사용자 등록
+    }
+
+    // 새로운 카카오 사용자 등록
+    @Transactional
+    public Member registerNewKakaoUser(KakaoUserInfoDTO kakaoUserInfo) {
+        // 우선 닉네임 없이 회원 생성 (ID를 얻기 위해)
+        Member member = Member.builder()
+                .socialId(kakaoUserInfo.getId())
+                .email(UUID.randomUUID() + "@socialUser.com")
+                .kakaoNickname(kakaoUserInfo.getNickname())
+                .role(Role.USER)
+                .profileImage(null)
+                .build();
+
+        return memberRepository.save(member);
+    }
+}
