+++ src/main/java/com/rhkr8521/spaceon/common/exception/NotFoundException.java
@@ -1,0 +1,13 @@
+package com.rhkr8521.spaceon.common.exception;
+
+import org.springframework.http.HttpStatus;
+
+public class NotFoundException extends BaseException{
+    public NotFoundException() {
+        super(HttpStatus.NOT_FOUND);
+    }
+
+    public NotFoundException(String message) {
+        super(HttpStatus.NOT_FOUND, message);
+    }
+}
