--- src/main/java/com/rhkr8521/zerocommission/api/member/dto/KakaoLoginRequestDTO.java
@@ -1,8 +1,0 @@
-package com.rhkr8521.zerocommission.api.member.dto;
-
-import lombok.Data;
-
-@Data
-public class KakaoLoginRequestDTO {
-    private String accessToken;
-}
