--- src/main/java/com/rhkr8521/zerocommission/api/member/repository/MemberRepository.java
@@ -1,17 +1,0 @@
-package com.rhkr8521.zerocommission.api.member.repository;
-
-import com.rhkr8521.zerocommission.api.member.entity.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
-    Optional<Member> findByEmail(String email);
-
-    Optional<Member> findByRefreshToken(String refreshToken);
-
-    Optional<Member> findBySocialId(String socialId);
-
-    boolean existsByNickname(String nickname);
-}
