--- src/main/java/com/rhkr8521/zerocommission/common/advice/ControllerExceptionAdvice.java
@@ -1,45 +1,0 @@
-package com.rhkr8521.zerocommission.common.advice;
-
-import java.util.Objects;
-
-import com.rhkr8521.zerocommission.common.exception.BaseException;
-import com.rhkr8521.zerocommission.common.response.ApiResponse;
-import com.rhkr8521.zerocommission.common.response.ErrorStatus;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.FieldError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.MissingServletRequestParameterException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class ControllerExceptionAdvice {
-
-    @ExceptionHandler(BaseException.class)
-    public ResponseEntity<ApiResponse> handleGlobalException(BaseException ex) {
-        return ResponseEntity.status(ex.getStatusCode())
-                .body(ApiResponse.fail(ex.getStatusCode(), ex.getResponseMessage()));
-    }
-
-    @ExceptionHandler(MissingServletRequestParameterException.class)
-    public ResponseEntity<ApiResponse> handleMissingParameter(MissingServletRequestParameterException ex) {
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-                .body(ApiResponse.fail(HttpStatus.BAD_REQUEST.value(),
-                        ErrorStatus.VALIDATION_REQUEST_MISSING_EXCEPTION.getMessage()));
-    }
-
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ResponseEntity<ApiResponse>  handleIllegalArgument(IllegalArgumentException ex) {
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-                .body(ApiResponse.fail(HttpStatus.BAD_REQUEST.value(), ex.getMessage()));
-    }
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    protected ResponseEntity<ApiResponse> handleMethodArgumentNotValidException(final MethodArgumentNotValidException e) {
-        FieldError fieldError = Objects.requireNonNull(e.getFieldError());
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-                .body(ApiResponse.fail(HttpStatus.BAD_REQUEST.value(),String.format("%s. (%s)", fieldError.getDefaultMessage(), fieldError.getField())));
-    }
-
-}
