--- src/main/java/com/rhkr8521/zerocommission/common/config/oauth2/OAuth2Config.java
@@ -1,107 +1,0 @@
-package com.rhkr8521.zerocommission.common.config.oauth2;
-
-import com.rhkr8521.zerocommission.api.member.service.MemberService;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.RequestEntity;
-import org.springframework.security.oauth2.client.InMemoryOAuth2AuthorizedClientService;
-import org.springframework.security.oauth2.client.OAuth2AuthorizedClientService;
-import org.springframework.security.oauth2.client.endpoint.OAuth2AuthorizationCodeGrantRequest;
-import org.springframework.security.oauth2.client.endpoint.OAuth2AuthorizationCodeGrantRequestEntityConverter;
-import org.springframework.security.oauth2.client.endpoint.DefaultAuthorizationCodeTokenResponseClient;
-import org.springframework.security.oauth2.client.registration.ClientRegistration;
-import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository;
-import org.springframework.security.oauth2.client.registration.InMemoryClientRegistrationRepository;
-import org.springframework.security.oauth2.core.ClientAuthenticationMethod;
-
-@Configuration
-@RequiredArgsConstructor
-@Getter
-public class OAuth2Config {
-
-    private final MemberService customOAuth2UserService;
-
-    @Value("${spring.security.oauth2.client.registration.kakao.client-id}")
-    private String clientId;
-
-    @Value("${spring.security.oauth2.client.registration.kakao.client-secret}")
-    private String clientSecret;
-
-    @Value("${spring.security.oauth2.client.registration.kakao.redirect-uri}")
-    private String redirectUri;
-
-    @Value("${spring.security.oauth2.client.registration.kakao.client-authentication-method}")
-    private String clientAuthenticationMethod;
-
-    @Value("${spring.security.oauth2.client.registration.kakao.authorization-grant-type}")
-    private String authorizationGrantType;
-
-    @Value("${spring.security.oauth2.client.registration.kakao.scope}")
-    private String scope;
-
-    @Value("${spring.security.oauth2.client.provider.kakao.authorization-uri}")
-    private String authorizationUri;
-
-    @Value("${spring.security.oauth2.client.provider.kakao.token-uri}")
-    private String tokenUri;
-
-    @Value("${spring.security.oauth2.client.provider.kakao.user-info-uri}")
-    private String userInfoUri;
-
-    @Value("${spring.security.oauth2.client.provider.kakao.user-name-attribute}")
-    private String userNameAttributeName;
-
-    @Bean
-    public ClientRegistrationRepository clientRegistrationRepository() {
-        return new InMemoryClientRegistrationRepository(this.kakaoClientRegistration());
-    }
-
-    private ClientRegistration kakaoClientRegistration() {
-        return ClientRegistration.withRegistrationId("kakao")
-                .clientId(clientId)
-                .clientSecret(clientSecret)
-                .clientAuthenticationMethod(ClientAuthenticationMethod.CLIENT_SECRET_POST)
-                .authorizationGrantType(new org.springframework.security.oauth2.core.AuthorizationGrantType(authorizationGrantType))
-                .redirectUri(redirectUri)
-                .scope(scope.trim().split("\\s*,\\s*"))
-                .authorizationUri(authorizationUri)
-                .tokenUri(tokenUri)
-                .userInfoUri(userInfoUri)
-                .userNameAttributeName(userNameAttributeName)
-                .clientName("Kakao")
-                .build();
-    }
-
-    @Bean
-    public OAuth2AuthorizedClientService authorizedClientService() {
-        return new InMemoryOAuth2AuthorizedClientService(clientRegistrationRepository());
-    }
-
-    @Bean
-    public DefaultAuthorizationCodeTokenResponseClient authorizationCodeTokenResponseClient() {
-        DefaultAuthorizationCodeTokenResponseClient client = new DefaultAuthorizationCodeTokenResponseClient();
-        client.setRequestEntityConverter(new CustomOAuth2AuthorizationRequestEntityConverter());
-        return client;
-    }
-
-    // CustomOAuth2AuthorizationRequestEntityConverter 클래스 추가
-    public class CustomOAuth2AuthorizationRequestEntityConverter extends OAuth2AuthorizationCodeGrantRequestEntityConverter {
-
-        @Override
-        public RequestEntity<?> convert(OAuth2AuthorizationCodeGrantRequest authorizationCodeGrantRequest) {
-            RequestEntity<?> requestEntity = super.convert(authorizationCodeGrantRequest);
-
-            if (authorizationCodeGrantRequest.getClientRegistration().getClientAuthenticationMethod() == ClientAuthenticationMethod.CLIENT_SECRET_POST) {
-                return RequestEntity
-                        .post(requestEntity.getUrl())
-                        .headers(requestEntity.getHeaders())
-                        .body(requestEntity.getBody());
-            }
-
-            return requestEntity;
-        }
-    }
-}
