--- src/main/java/com/rhkr8521/zerocommission/common/entity/BaseTimeEntity.java
@@ -1,24 +1,0 @@
-package com.rhkr8521.zerocommission.common.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import java.time.LocalDateTime;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-@Getter
-public abstract class BaseTimeEntity {
-
-    @CreatedDate
-    @Column(updatable = false)
-    protected LocalDateTime createdAt;
-
-    @LastModifiedDate
-    protected LocalDateTime updatedAt;
-}
-
