--- src/main/java/com/rhkr8521/zerocommission/common/exception/BadRequestException.java
@@ -1,13 +1,0 @@
-package com.rhkr8521.zerocommission.common.exception;
-
-import org.springframework.http.HttpStatus;
-
-public class BadRequestException extends BaseException{
-    public BadRequestException() {
-        super(HttpStatus.BAD_REQUEST);
-    }
-
-    public BadRequestException(String message) {
-        super(HttpStatus.BAD_REQUEST, message);
-    }
-}
