--- src/main/java/com/rhkr8521/zerocommission/common/exception/BaseException.java
@@ -1,30 +1,0 @@
-package com.rhkr8521.zerocommission.common.exception;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.http.HttpStatus;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class BaseException extends RuntimeException {
-
-    HttpStatus statusCode;
-    String responseMessage;
-
-    public BaseException(HttpStatus statusCode) {
-        super();
-        this.statusCode = statusCode;
-    }
-
-    public BaseException(HttpStatus statusCode, String responseMessage) {
-        super(responseMessage);
-        this.statusCode = statusCode;
-        this.responseMessage = responseMessage;
-    }
-
-    public int getStatusCode() {
-        return this.statusCode.value();
-    }
-}
-
