--- src/main/java/com/rhkr8521/zerocommission/common/exception/InternalServerException.java
@@ -1,13 +1,0 @@
-package com.rhkr8521.zerocommission.common.exception;
-
-import org.springframework.http.HttpStatus;
-
-public class InternalServerException extends BaseException{
-    public InternalServerException() {
-        super(HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-
-    public InternalServerException(String message) {
-        super(HttpStatus.INTERNAL_SERVER_ERROR, message);
-    }
-}
