--- src/main/java/com/rhkr8521/zerocommission/common/exception/UnauthorizedException.java
@@ -1,13 +1,0 @@
-package com.rhkr8521.zerocommission.common.exception;
-
-import org.springframework.http.HttpStatus;
-
-public class UnauthorizedException extends BaseException{
-    public UnauthorizedException() {
-        super(HttpStatus.UNAUTHORIZED);
-    }
-
-    public UnauthorizedException(String message) {
-        super(HttpStatus.UNAUTHORIZED, message);
-    }
-}
