+++ src/test/java/com/rhkr8521/spaceon/SpaceonApplicationTests.java
@@ -1,0 +1,13 @@
+package com.rhkr8521.spaceon;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SpaceonApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
