--- src/test/java/com/rhkr8521/zerocommission/ZerocommissionApplicationTests.java
@@ -1,13 +1,0 @@
-package com.rhkr8521.zerocommission;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ZerocommissionApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
